+++ changelog.md
@@ -15,6 +15,7 @@ Version 20240314.9000+dev:
   * Add `[unselect_unit]` WML action.
   * Add `[open_doors]` WML action to switch gate terrain to its open version.
   * Add `[variable_exists]` WML conditional tag.
+  * Add `[list_foreach]` WML action tag.
 
 * Terrains:
   * Removed obsolete Earthy Rockbound Cave terrain.